Migrate GFTT to features module.
