+++ modules/cudafeatures2d/src/cuda/corners.cu
