+++ modules/cudaoptflow/perf/perf_precomp.hpp
@@ -47,6 +47,7 @@
 
 #include "opencv2/cudaoptflow.hpp"
 #include "opencv2/cudaarithm.hpp"
+#include "opencv2/cudafeatures2d.hpp"
 #include "opencv2/video.hpp"
 #include "opencv2/optflow.hpp"
