+++ modules/gapi/include/opencv2/gapi/imgproc.hpp
@@ -9,6 +9,7 @@
 #define OPENCV_GAPI_IMGPROC_HPP
 
 #include <opencv2/imgproc.hpp>
+#include <opencv2/features.hpp>
 
 #include <utility> // std::tuple
