+++ modules/gapi/src/precomp.hpp
@@ -12,6 +12,7 @@
 #  include <opencv2/core.hpp>
 #  include <opencv2/core/utils/configuration.private.hpp>
 #  include <opencv2/imgproc.hpp>
+#  include <opencv2/features.hpp>
 #  include <opencv2/gapi/core.hpp>
 #  include <opencv2/gapi/imgproc.hpp>
 #endif //  !defined(GAPI_STANDALONE)
