+++ modules/xstereo/src/precomp.hpp
@@ -44,6 +44,7 @@
 
 #include "opencv2/core.hpp"
 #include "opencv2/imgproc.hpp"
+#include "opencv2/features.hpp"
 #include "opencv2/stereo.hpp"
 
 #include "opencv2/xstereo.hpp"
